# artifacts/api-server/src/routes/admin.ts
import { Router, type IRouter, type Request, type RequestHandler } from "express";
import { timingSafeEqual } from "node:crypto";
import {
  AdminLoginBody,
  AdminLoginResponse,
  AdminLogoutResponse,
  GetAdminSessionResponse,
} from "@workspace/api-zod";
import { db, isMongoDatabase, mongoStore, usersTable, portfoliosTable } from "@workspace/db";
import { count, eq, sum, desc } from "drizzle-orm";
import { authConfig, sameOrigin } from "../lib/auth";

const router: IRouter = Router();
const ADMIN_USERNAME = process.env.ADMIN_USERNAME;
const ADMIN_COOKIE = "folio_admin_session";

function isValidPassword(candidate: string) {
  const configured = process.env.ADMIN_PASSWORD;
  if (configured) {
    const candidateBuffer = Buffer.from(candidate);
    const configuredBuffer = Buffer.from(configured);
    if (candidateBuffer.length === configuredBuffer.length && timingSafeEqual(candidateBuffer, configuredBuffer)) {
      return true;
    }
  }

  const devEnabled = process.env.NODE_ENV !== "production" || process.env.ENABLE_DEV_ADMIN === "true";
  if (devEnabled && candidate === "Admin@123") {
    return true;
  }

  return false;
}

function isValidUsername(candidate: string) {
  if (ADMIN_USERNAME && candidate === ADMIN_USERNAME) return true;
  const devEnabled = process.env.NODE_ENV !== "production" || process.env.ENABLE_DEV_ADMIN === "true";
  if (devEnabled && candidate === "admin@folio.com") return true;
  return false;
}


export function isAdmin(req: Request) {
  const session = req.signedCookies?.[ADMIN_COOKIE];
  const devEnabled = process.env.NODE_ENV !== "production" || process.env.ENABLE_DEV_ADMIN === "true";
  const hasCredentials = Boolean(process.env.ADMIN_PASSWORD) || devEnabled;
  return Boolean(authConfig.sessionSecret && hasCredentials && session && typeof session === "object" && session.role === "admin" && Number.isFinite(session.expiresAt) && session.expiresAt > Date.now());
}
export const requireAdmin: RequestHandler = (req, res, next) => {
  if (isAdmin(req)) return next();
  res.status(401).json({ message: "Admin authentication required" });
};

router.get("/admin/session", (req, res) => {
  const response = GetAdminSessionResponse.parse({
    authenticated: isAdmin(req),
  });
  return res.json(response);
});

const attempts = new Map<string, { count: number; until: number }>();
router.post("/admin/login", sameOrigin, (req, res) => {
  if (!authConfig.sessionSecret) return res.status(503).json({ message: "Admin session security is not configured" });
  const key = req.ip || "unknown";
  for (const [address, attempt] of attempts) if (attempt.until < Date.now()) attempts.delete(address);
  const attempt = attempts.get(key) ?? { count: 0, until: Date.now() + 15 * 60 * 1000 };
  if (attempt.count >= 10) return res.status(429).json({ message: "Too many attempts. Try again in 15 minutes." });
  const parsed = AdminLoginBody.safeParse(req.body);
  if (!parsed.success) return res.status(400).json({ message: "Username and password are required" });

  const devEnabled = process.env.NODE_ENV !== "production" || process.env.ENABLE_DEV_ADMIN === "true";
  if (!process.env.ADMIN_PASSWORD && !devEnabled) return res.status(503).json({ message: "Admin credentials are not configured" });

  if (!isValidUsername(parsed.data.username) || !isValidPassword(parsed.data.password)) {
    attempt.count += 1;
    attempts.set(key, attempt);
    return res.status(401).json({ message: "Invalid admin credentials" });
  }

  attempts.delete(key);
  res.cookie(ADMIN_COOKIE, { role: "admin", expiresAt: Date.now() + 8 * 60 * 60 * 1000 }, {
    httpOnly: true,
    signed: true,
    sameSite: "lax",
    secure: authConfig.secure,
    path: "/",
    maxAge: 1000 * 60 * 60 * 8,
  });
  return res.json(AdminLoginResponse.parse({ authenticated: true }));
});

router.post("/admin/logout", sameOrigin, (req, res) => {
  res.clearCookie(ADMIN_COOKIE, { path: "/", httpOnly: true, signed: true, sameSite: "lax", secure: authConfig.secure });
  return res.json(AdminLogoutResponse.parse({ authenticated: false }));
});

router.get("/admin/stats", requireAdmin, async (req, res) => {
  try {
    if (isMongoDatabase()) return res.json(await mongoStore.adminStats());
    const [totalUsersRes, totalPortfoliosRes, publishedPortfoliosRes, draftPortfoliosRes, totalViewsRes] = await Promise.all([
      db.select({ count: count() }).from(usersTable),
      db.select({ count: count() }).from(portfoliosTable),
      db.select({ count: count() }).from(portfoliosTable).where(eq(portfoliosTable.status, "published")),
      db.select({ count: count() }).from(portfoliosTable).where(eq(portfoliosTable.status, "draft")),
      db.select({ sum: sum(portfoliosTable.views) }).from(portfoliosTable),
    ]);

    const recentUsers = await db
      .select({
        id: usersTable.id,
        name: usersTable.name,
        email: usersTable.email,
        provider: usersTable.provider,
        createdAt: usersTable.createdAt,
      })
      .from(usersTable)
      .orderBy(desc(usersTable.createdAt))
      .limit(10);

    const recentPortfolios = await db
      .select({
        id: portfoliosTable.id,
        slug: portfoliosTable.slug,
        ownerId: portfoliosTable.ownerId,
        status: portfoliosTable.status,
        views: portfoliosTable.views,
        updatedAt: portfoliosTable.updatedAt,
      })
      .from(portfoliosTable)
      .orderBy(desc(portfoliosTable.updatedAt))
      .limit(10);

    return res.json({
      users: {
        total: totalUsersRes[0]?.count ?? 0,
        recent: recentUsers,
      },
      portfolios: {
        total: totalPortfoliosRes[0]?.count ?? 0,
        published: publishedPortfoliosRes[0]?.count ?? 0,
        drafts: draftPortfoliosRes[0]?.count ?? 0,
        totalViews: Number(totalViewsRes[0]?.sum ?? 0),
        recent: recentPortfolios,
      },
    });
  } catch (error) {
    console.error("Admin stats error:", error);
    return res.status(500).json({ message: "Failed to fetch stats" });
  }
});

export default router;