# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: portfolios
    description: Portfolio creation and publishing
  - name: templates
    description: Portfolio template marketplace
  - name: resume
    description: Resume parsing operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /templates:
    get:
      operationId: listTemplates
      tags: [templates]
      summary: List portfolio templates
      responses:
        "200":
          description: Available templates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PortfolioTemplate"
  /portfolios:
    get:
      operationId: getCurrentPortfolio
      tags: [portfolios]
      summary: Get the current portfolio
      responses:
        "200":
          description: Current portfolio
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Portfolio"
    post:
      operationId: createPortfolio
      tags: [portfolios]
      summary: Create a portfolio draft
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PortfolioInput"
      responses:
        "201":
          description: Created portfolio
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Portfolio"
  /portfolios/{id}:
    patch:
      operationId: updatePortfolio
      tags: [portfolios]
      summary: Update portfolio content
      parameters:
        - $ref: "#/components/parameters/PortfolioId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PortfolioUpdate"
      responses:
        "200":
          description: Updated portfolio
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Portfolio"
    get:
      operationId: getPortfolio
      tags: [portfolios]
      summary: Get a portfolio draft
      parameters:
        - $ref: "#/components/parameters/PortfolioId"
      responses:
        "200":
          description: Portfolio
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Portfolio"
  /portfolios/{id}/publish:
    post:
      operationId: publishPortfolio
      tags: [portfolios]
      summary: Publish a portfolio
      parameters:
        - $ref: "#/components/parameters/PortfolioId"
      responses:
        "200":
          description: Published portfolio
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Portfolio"
  /portfolios/{id}/unpublish:
    post:
      operationId: unpublishPortfolio
      tags: [portfolios]
      summary: Unpublish a portfolio
      parameters:
        - $ref: "#/components/parameters/PortfolioId"
      responses:
        "200":
          description: Unpublished portfolio
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Portfolio"
  /public/portfolios:
    get:
      operationId: listPublishedPortfolios
      tags: [portfolios]
      summary: List published portfolio slugs for sitemap generation
      responses:
        "200":
          description: Published portfolios
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  required: [slug, updatedAt]
                  properties:
                    slug: { type: string }
                    updatedAt: { type: string }
                    publishedAt: { type: ["string", "null"] }
  /public/portfolios/{slug}:
    get:
      operationId: getPublicPortfolio
      tags: [portfolios]
      summary: Read a published portfolio without incrementing views
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Public portfolio
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Portfolio"
        "404":
          description: Portfolio not found
  /public/portfolios/{slug}/views:
    post:
      operationId: recordPublicPortfolioView
      tags: [portfolios]
      summary: Record a public portfolio page view
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: View recorded
        "404":
          description: Portfolio not found
  /admin/session:
    get:
      operationId: getAdminSession
      tags: [admin]
      summary: Check the current admin session
      responses:
        "200":
          description: Admin session state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Sign in to the admin workspace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginInput"
      responses:
        "200":
          description: Authenticated admin session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "401":
          description: Invalid admin credentials
        "503":
          description: Admin credentials are not configured
  /admin/logout:
    post:
      operationId: adminLogout
      tags: [admin]
      summary: End the current admin session
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Get admin dashboard statistics
      responses:
        "200":
          description: Admin stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStatsResponse"
  /resume/parse:
    post:
      operationId: parseResume
      tags: [resume]
      summary: Extract portfolio data from a resume
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/ResumeParseInput"
      responses:
        "200":
          description: Extracted information awaiting review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResumeExtraction"
components:
  parameters:
    PortfolioId:
      name: id
      in: path
      required: true
      schema:
        type: string
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
        geminiConfigured:
          type: boolean
      required:
        - status
        - geminiConfigured
    AdminLoginInput:
      type: object
      required: [username, password]
      properties:
        username: { type: string, minLength: 1 }
        password: { type: string, minLength: 1 }
    AdminSession:
      type: object
      required: [authenticated]
      properties:
        authenticated: { type: boolean }
    AdminStatsResponse:
      type: object
      required: [users, portfolios]
      properties:
        users:
          type: object
          required: [total, recent]
          properties:
            total: { type: integer }
            recent:
              type: array
              items:
                type: object
                required: [id, name, provider, createdAt]
                properties:
                  id: { type: string }
                  name: { type: string }
                  email: { type: ["string", "null"] }
                  provider: { type: string }
                  createdAt: { type: string }
        portfolios:
          type: object
          required: [total, published, drafts, totalViews, recent]
          properties:
            total: { type: integer }
            published: { type: integer }
            drafts: { type: integer }
            totalViews: { type: integer }
            recent:
              type: array
              items:
                type: object
                required: [id, slug, ownerId, status, views, updatedAt]
                properties:
                  id: { type: integer }
                  slug: { type: string }
                  ownerId: { type: string }
                  status: { type: string }
                  views: { type: integer }
                  updatedAt: { type: string }
    PortfolioTemplate:
      type: object
      required: [id, name, category, description, premium, accent, layout]
      properties:
        id: { type: string }
        name: { type: string }
        category: { type: string }
        description: { type: string }
        premium: { type: boolean }
        accent: { type: string }
        layout: { type: string }
        recommendedFor:
          type: array
          items: { type: string }
    PersonalInfo:
      type: object
      required: [name, headline, email, phone, location, summary]
      properties:
        name: { type: string }
        headline: { type: string }
        email: { type: string }
        phone: { type: string }
        location: { type: string }
        summary: { type: string }
        avatar: { type: ["string", "null"] }
    Experience:
      type: object
      required: [id, role, company, period, description]
      properties:
        id: { type: string }
        role: { type: string }
        company: { type: string }
        period: { type: string }
        description: { type: string }
    Education:
      type: object
      required: [id, school, degree, period]
      properties:
        id: { type: string }
        school: { type: string }
        degree: { type: string }
        period: { type: string }
    Project:
      type: object
      required: [id, name, description, technologies, githubUrl, liveUrl]
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string }
        technologies:
          type: array
          items: { type: string }
        githubUrl: { type: string }
        liveUrl: { type: string }
        image: { type: ["string", "null"] }
    SocialLinks:
      type: object
      required: [github, linkedin, twitter, website]
      properties:
        github: { type: string }
        linkedin: { type: string }
        twitter: { type: string }
        website: { type: string }
    PortfolioContent:
      type: object
      required: [personalInfo, experience, education, skills, projects, socialLinks]
      properties:
        personalInfo:
          $ref: "#/components/schemas/PersonalInfo"
        experience:
          type: array
          items: { $ref: "#/components/schemas/Experience" }
        education:
          type: array
          items: { $ref: "#/components/schemas/Education" }
        skills:
          type: array
          items: { type: string }
        projects:
          type: array
          items: { $ref: "#/components/schemas/Project" }
        socialLinks:
          $ref: "#/components/schemas/SocialLinks"
    Portfolio:
      type: object
      required: [id, slug, profession, purpose, templateId, status, content, views, updatedAt]
      properties:
        id: { type: string }
        slug: { type: string }
        profession: { type: string }
        purpose: { type: string }
        templateId: { type: string }
        status: { type: string, enum: [draft, published] }
        content:
          $ref: "#/components/schemas/PortfolioContent"
        views: { type: integer }
        updatedAt: { type: string }
        publishedAt: { type: ["string", "null"] }
    PortfolioInput:
      type: object
      required: [profession, purpose, templateId, content]
      properties:
        profession: { type: string }
        purpose: { type: string }
        templateId: { type: string }
        content:
          $ref: "#/components/schemas/PortfolioContent"
    PortfolioUpdate:
      type: object
      properties:
        profession: { type: string }
        purpose: { type: string }
        templateId: { type: string }
        content:
          $ref: "#/components/schemas/PortfolioContent"
    ResumeParseInput:
      type: object
      required: [file]
      properties:
        file:
          type: string
          format: binary
    ResumeExtraction:
      type: object
      required: [fileName, extracted, warnings]
      properties:
        fileName: { type: string }
        extracted:
          $ref: "#/components/schemas/PortfolioContent"
        warnings:
          type: array
          items: { type: string }